+++ client/src/App.js
@@ -1,15 +1,13 @@
-import React, { Component } from "react";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import Recipes from "./components/Recipes";
-import NotFound from "./components/NotFound";
-import Nav from "./components/Nav";
+import React, { Component } from 'react';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import Recipes from './components/Recipes';
+import NotFound from './components/NotFound';
 
 class App extends Component {
   render() {
     return (
       <Router>
         <div className="App">
-          <Nav />
           <Switch>
             <Route exact path="/recipes" component={Recipes} />
             <Route component={NotFound} />
