+++ client/src/components/Grid/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './style.css';
+
+function GridContainer(props) {
+  return <div className="grid-container" {...props} />;
+}
+
+export default GridContainer;
