+++ home.css
@@ -1,0 +1,53 @@
+html {
+  position: relative;
+  min-height: 100%;
+}
+
+body {
+  margin-bottom: 60px; /* Margin bottom by footer height */
+  background-image: url('./client/public/images/home.jpg');
+  -webkit-background-size: cover;
+  -moz-background-size: cover;
+  -o-background-size: cover;
+  background-size: cover;
+  font-family: 'Roboto', sans-serif;
+  color: white;
+}
+
+.main-wrapper {
+  margin: auto;
+  text-align: right;
+}
+
+.grid-container {
+  text-align: left;
+  color: black;
+  margin: 10px;
+  display: grid;
+  grid-template-columns: 700px 1fr;
+  padding-bottom: 100px;
+  /* background: rgba(255, 255, 255, 0.5); */
+  padding: 20px;
+}
+.newsfeed,
+.snapshot,
+.workouts,
+.food {
+  background: rgba(255, 255, 255, 0.5);
+  border-radius: 2px;
+  margin-bottom: 2px;
+}
+
+.snapshot img {
+  width: 100%;
+  border-radius: 2px;
+}
+
+.footer {
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  height: 30px; /* Set the fixed height of the footer here */
+  line-height: 30px; /* Vertically center the text there */
+  text-align: center;
+}
