+++ login.css
@@ -1,0 +1,45 @@
+html {
+  position: relative;
+  min-height: 100%;
+}
+
+body {
+  margin-bottom: 60px; /* Margin bottom by footer height */
+  background-image: url('./client/public/images/gymwall.jpg');
+  -webkit-background-size: cover;
+  -moz-background-size: cover;
+  -o-background-size: cover;
+  background-size: cover;
+  font-family: 'Roboto', sans-serif;
+  color: white;
+}
+
+.login-container {
+  margin: 50px;
+  display: grid;
+  grid-gap: 10px;
+  grid-template-columns: 370px;
+  padding-bottom: 100px;
+}
+
+.login-container img {
+  width: 50px;
+  background-color: transparent;
+}
+
+.grid-container {
+  margin: 0 50px;
+  display: grid;
+  grid-gap: 10px;
+  grid-template-columns: 1fr 1fr;
+  padding-bottom: 100px;
+}
+
+.footer {
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  height: 30px; /* Set the fixed height of the footer here */
+  line-height: 30px; /* Vertically center the text there */
+  text-align: center;
+}
