+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "fitness-warrior",
+  "name": "SpartanFit",
   "version": "1.0.0",
-  "description": "Mern Demo",
+  "description": "SpartanFit",
   "main": "server.js",
   "scripts": {
     "start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
